requirements.txt
fastapi==0.112.1
uvicorn==0.30.6
python-dotenv==1.0.0
supabase
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
pytest==8.3.2
httpx==0.27.2
bcrypt==4.0.1
email_validator==2.2.0
pydantic==2.8.2
pydantic_core==2.20.1
pydantic-settings